refactor(api): split 1944-line marketing_templates.py into per-category modules
Extract shared field helpers to _fields.py, category metadata to _categories.py,
and the 52 template definitions into 7 per-category modules under
src/templates/marketing/. marketing_templates.py is now a 42-line assembler that
re-exports MARKETING_TEMPLATES and TEMPLATE_CATEGORIES, so all import paths are
unchanged. Verified behavior-preserving: the assembled registry hashes identically
to the pre-refactor version (52 templates) and the full backend suite passes.
