+++ apps/api/src/templates/marketing/__init__.py
@@ -1,0 +1,1 @@
+"""Per-category marketing template lists."""
